FLGO-2 setup modules
