+++ config/db.go
@@ -1,0 +1,9 @@
+package config
+
+import (
+	"github.com/jinzhu/gorm"
+)
+
+type FLDB struct {
+	*gorm.DB
+}
