+++ go.mod
@@ -1,0 +1,8 @@
+module fantlab
+
+go 1.12
+
+require (
+	github.com/gin-gonic/gin v1.4.0
+	github.com/jinzhu/gorm v1.9.8
+)
