+++ main.go
@@ -1,15 +1,17 @@
 package main
 
 import (
+	"fantlab/config"
+	"fantlab/routing"
 	"log"
 
-	"github.com/gin-gonic/gin"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/mysql"
 )
 
 func main() {
-	db, err := gorm.Open("mysql", "root:root@/fantlab?charset=utf8&parseTime=True&loc=Europe%2FMoscow")
+	// db, err := gorm.Open("mysql", "root:root@/fantlab?charset=utf8&parseTime=True&loc=Europe%2FMoscow")
+	db, err := gorm.Open("mysql", "root@/fl?charset=utf8&parseTime=True&loc=Local")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -21,16 +23,9 @@ func main() {
 	}()
 
 	db.LogMode(true)
-	fdb := &FDB{db}
+	fldb := &config.FLDB{db}
 
-	router := gin.New()
-	router.Use(gin.Logger(), gin.Recovery())
-
-	controller := NewController(fdb)
-
-	router.GET("/forums", controller.ShowForums)
-	router.GET("/forums/:id", controller.ShowForumTopics)
-	router.GET("/topics/:id", controller.ShowTopicMessages)
+	router := routing.SetupWith(fldb)
 
 	if err := router.Run(":4242"); err != nil {
 		log.Fatal(err)
